requirements.txt
# configuración de dependencias para pelear contra render que no quiere correr la pagina

Flask==2.3.3
Flask-SocketIO==5.5.1
python-socketio==5.13.0
python-engineio==4.12.2
eventlet==0.33.3
# utilidades (no estrictamente necesarias pero ayudan)
Jinja2==3.1.6
Werkzeug==3.1.3
MarkupSafe==3.0.2
itsdangerous==2.2.0
click==8.2.1
blinker==1.9.0
greenlet==3.2.4
